+++ Blurred-ios/OtherFollowingCell.swift
@@ -1,23 +1,20 @@
 //
-//  Follower.swift
+//  OtherFollowingCell.swift
 //  Blurred-ios
 //
-//  Created by Martin Velev on 5/9/20.
+//  Created by Martin Velev on 5/19/20.
 //  Copyright © 2020 BlurrMC. All rights reserved.
 //
 
 import UIKit
 
-class FollowerCell: UITableViewCell {
+class OtherFollowingCell: UITableViewCell {
 
     override func awakeFromNib() {
         super.awakeFromNib()
         // Initialization code
     }
-    
-    @IBOutlet weak var followerAvatar: UIImageView!
-    @IBOutlet weak var followerName: UILabel!
-    @IBOutlet weak var followerUsername: UILabel!
+
     override func setSelected(_ selected: Bool, animated: Bool) {
         super.setSelected(selected, animated: animated)